requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-multipart==0.0.6
pandas==2.1.3
